fix(applicationset): ApplicationSets with rolling sync stuck in Pending
In issue https://github.com/argoproj/argo-cd/issues/19535 we have discovered a bug where ApplicationSet Progressive Sync feature gets stuck and it never recovers and get stuck forever unless you manually delete the existing `applicationsStatus` of the ApplicationSet affected or trigger a manual sync.

When the ApplicationSet is performing a progressive sync, the apps in the step being synced get the status `Pending` in the ApplicationSet `applicationStatus `. This means that the apps are gonna be synced and is waiting for the sync to start progressing.

Let's set an example, applicationset generates 3 applications. In the
initial moment applicationset points to commit A

applicationset will generate those 3 applications and start the
progressive sync, then application 2 is in Pending status, the
applicationset status for application 2 is marked to targetrevision for
app2 to A

At this moment in time applicationset gets updated to point to commit B,
since app2 is in pending state the progressive sync allows it to the app
to be synced and hence the app2 is synced to commit B

since applicationset targetrevision for app2 expects to be A but it's B
will never move app2 from 'Pending' to 'Progressing' state.

[Here](https://github.com/argoproj/argo-cd/blob/master/applicationset/controllers/applicationset_controller.go#L1078) is the logic that performs this check.

This new check was introduced in ArgoCD 2.12 causing this bug when a progressive sync is already being performed.

- Because first it will get the `applicationStatus` from the existing applicationStatus which is the one that has the [old revision](https://github.com/argoproj/argo-cd/blob/master/applicationset/controllers/applicationset_controller.go#L1045)
- And since the app inside the ApplicationSet `applicationStatus` is in "Pending" the revision is never updated when it enters the [if statement](https://github.com/argoproj/argo-cd/blob/master/applicationset/controllers/applicationset_controller.go#L1069-L1092) (see how currentAppStatus.TargetRevision never will be updated)

This means that the ApplicationSet will always think that the app is not being synced to the latest version and never progress, but in reality the app is actually in a later version but tha ApplicationSet never updated it in the apps that are in "Pending".

- This PR fixes this bug changing the logic that checks when an
  applications needs to be moved from Pending to Progressing, instead of
rely on the targetrevision we actually rely just in the application
being synced to move it. This also don't introduce a prior bug where it
was cheched that the application was synced in a certain moment in time
to ensure it was triggered by the applicationset controller.

- Note that if someone manually sync one application of the
  applicationset while it's being progresively synced after merging this
PR the applciationset controller will continue the rollout

- Ensure that a certain revision is applied orderly in all applications
  generated from the applicationset it's certainly possible that a given
application can be synced in a different revision than the one
explicitly set in the appset

Co-authored-by: Thibault Jamet <tjamet@user.noreply.github.com>
Co-authored-by: Carlos Rejano <carlosrejanoromeu@gmail.com>